requirements.txt
APScheduler==3.7.0
asttokens==2.0.4
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
executing==0.5.4
fire==0.4.0
importlib-metadata==3.7.3
iniconfig==1.1.1
jsons==1.4.0
jsonschema==3.2.0
kiwisolver==1.3.1
littleutils==0.2.2
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
retrying==1.3.3
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sorcery==0.2.1
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
typish==1.9.1
tzlocal==2.1
uuid==1.30
watchdog==2.0.2
wrapt==1.12.1
zipp==3.4.1
